refactor photogrid
