+++ src/lib/album.ts
@@ -1,82 +1,102 @@
 import { writable } from 'svelte/store';
 import { set as idbSet, get as idbGet } from 'idb-keyval';
-
+import type { Photo } from './photos';
+import type { AlbumProps } from 'lucide-svelte/icons/album';
 
 export const requestedAlbumID = writable<string>('');
 
+export type Album = {
+	id: string;
+	title: string;
+	images: Photo['id'][];
+};
+
 export function requestAlbum(albumId: string) {
-    requestedAlbumID.set(albumId);
+	requestedAlbumID.set(albumId);
 }
 
-// Definiere das Album-Store
-export const albumStore = writable([]);
+// Define the Album Store
+export const albumStore = (() => {
+	const { subscribe, set, update } = writable<Album[]>([]);
 
-// Funktion zum Laden der Alben aus der IndexedDB
-export async function loadAlbums() {
-    const albums = await idbGet('albums');
-    if (albums) {
-        albumStore.set(albums);
-    }
-}
+	// Function to load the albums from IndexedDB
+	async function loadAlbums() {
+		const albums: Album[] = (await idbGet('albums')) ?? [];
+		if (albums) {
+			set(albums);
+		}
+	}
 
-// Funktion zum Hinzufügen eines Albums
-export function addAlbum(newAlbum) {
-    albumStore.update(albums => {
-        const updatedAlbums = [...albums, newAlbum];
-        idbSet('albums', updatedAlbums);
-        return updatedAlbums;
-    });
-}
+	// Function to add an album
+	function addAlbum(newAlbum: Album) {
+		update((albums) => {
+			const updatedAlbums = [...albums, newAlbum];
+			idbSet('albums', updatedAlbums);
+			return updatedAlbums;
+		});
+	}
 
-// Funktion zum Entfernen eines Albums
-export function removeAlbum(albumId) {
-    albumStore.update(albums => {
-        const updatedAlbums = albums.filter(album => album.id !== albumId);
-        idbSet('albums', updatedAlbums);
-        return updatedAlbums;
-    });
-}
+	// Function to remove an album
+	function removeAlbum(albumId: Album['id']) {
+		update((albums) => {
+			const updatedAlbums = albums.filter((album) => album.id !== albumId);
+			idbSet('albums', updatedAlbums);
+			return updatedAlbums;
+		});
+	}
 
-// Funktion zum Hinzufügen eines Bildes zu einem Album
-export function addImageToAlbum(albumId, newImage) {
-    albumStore.update(albums => {
-        const updatedAlbums = albums.map(album => {
-            if (album.id === albumId) {
-                return { ...album, images: [...album.images, newImage] };
-            }
-            return album;
-        });
-        idbSet('albums', updatedAlbums);
-        return updatedAlbums;
-    });
-}
+	// Function to add an image to an album
+	function addImageToAlbum(albumId: Album['id'], newImage: Photo['url']) {
+		update((albums) => {
+			const updatedAlbums = albums.map((album) => {
+				if (album.id === albumId) {
+					return { ...album, images: [...album.images, newImage] };
+				}
+				return album;
+			});
+			idbSet('albums', updatedAlbums);
+			return updatedAlbums;
+		});
+	}
 
-// Funktion zum Entfernen eines Bildes aus einem Album
-export function removeImageFromAlbum(albumId, imageId) {
-    albumStore.update(albums => {
-        const updatedAlbums = albums.map(album => {
-            if (album.id === albumId) {
-                return {
-                    ...album,
-                    images: album.images.filter(image => image.id !== imageId),
-                };
-            }
-            return album;
-        });
-        idbSet('albums', updatedAlbums);
-        return updatedAlbums;
-    });
-}
+	// Function to remove an image from an album
+	function removeImageFromAlbum(albumId: Album['id'], imageUrl: Photo['url']) {
+		update((albums) => {
+			const updatedAlbums = albums.map((album) => {
+				if (album.id === albumId) {
+					return {
+						...album,
+						images: album.images.filter((image) => image !== imageUrl)
+					};
+				}
+				return album;
+			});
+			idbSet('albums', updatedAlbums);
+			return updatedAlbums;
+		});
+	}
 
-export function updateAlbumTitle(albumId, newTitle) {
-    albumStore.update(albums => {
-        const updatedAlbums = albums.map(album => {
-            if (album.id === albumId) {
-                return { ...album, title: newTitle };
-            }
-            return album;
-        });
-        idbSet('albums', updatedAlbums);
-        return updatedAlbums;
-    });
-}
+	// Function to update the title of an album
+	function updateAlbumTitle(albumId: Album['id'], newTitle: Album['title']) {
+		update((albums) => {
+			const updatedAlbums = albums.map((album) => {
+				if (album.id === albumId) {
+					return { ...album, title: newTitle };
+				}
+				return album;
+			});
+			idbSet('albums', updatedAlbums);
+			return updatedAlbums;
+		});
+	}
+
+	return {
+		subscribe,
+		loadAlbums,
+		addAlbum,
+		removeAlbum,
+		addImageToAlbum,
+		removeImageFromAlbum,
+		updateAlbumTitle
+	};
+})();
