+++ src/lib/uistore.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export let open = writable(false);
