+++ svelte.config.js
@@ -14,7 +14,7 @@ const config = {
 			precompress: false,
 			strict: true
 		})
-	}
+	},
 };
 
 export default config;
